+++ python/solver.py
@@ -510,7 +510,7 @@ class ModeSolver(object):
         return update_brd(brd, freqs, [])
 
     def output_band_range_data(self, br_data):
-        if verbosity >= 2:
+        if verbosity.mpb >= 1:
             for tup, band in zip(br_data, range(1, len(br_data) + 1)):
                 fmt = "Band {} range: {} at {} to {} at {}"
                 min_band, max_band = tup
@@ -537,7 +537,7 @@ class ModeSolver(object):
                 else:
                     gap_size = ((200 * (br_rest_min_f - br_cur_max_f)) /
                                 (br_rest_min_f + br_cur_max_f))
-                    if verbosity >= 2:
+                    if verbosity.mpb >= 1:
                         fmt = "Gap from band {} ({}) to band {} ({}), {}%"
                         print(fmt.format(i, br_cur_max_f, i + 1, br_rest_min_f, gap_size))
                     return ogaps(br_rest[0], br_rest[1:], i + 1,
@@ -642,7 +642,7 @@ class ModeSolver(object):
             self.freqs[curfield_band - 1]
         )
         fname = self._create_fname(fname, fname_prefix, True)
-        if verbosity >= 2:
+        if verbosity.mpb >= 1:
             print("Outputting complex scalar field to {}...".format(fname))
 
         with h5py.File(fname, 'w') as f:
@@ -676,7 +676,7 @@ class ModeSolver(object):
         )
 
         fname = self._create_fname(fname, fname_prefix, True)
-        if verbosity >= 2:
+        if verbosity.mpb >= 1:
             print("Outputting fields to {}...".format(fname))
 
         with h5py.File(fname, 'w') as f:
@@ -721,7 +721,7 @@ class ModeSolver(object):
 
         parity_suffix = False if curfield_type in 'mn' else True
         fname = self._create_fname(fname, fname_prefix, parity_suffix)
-        if verbosity >= 2:
+        if verbosity.mpb >= 1:
             print("Outputting {}...".format(fname))
 
         with h5py.File(fname, 'w') as f:
@@ -810,7 +810,7 @@ class ModeSolver(object):
 
     def display_kpoint_data(self, name, data):
         k_index = self.mode_solver.get_kpoint_index()
-        if verbosity >= 2:
+        if verbosity.mpb >= 1:
             print("{}{}:, {}".format(self.parity, name, k_index), end='')
 
             for d in data:
@@ -826,7 +826,7 @@ class ModeSolver(object):
         min_iters = min(self.eigensolver_iters)
         max_iters = max(self.eigensolver_iters)
         mean_iters = np.mean(self.eigensolver_iters)
-        if verbosity >= 2:
+        if verbosity.mpb >= 1:
             fmt = "eigensolver iterations for {} kpoints: {}-{}, mean = {}"
             print(fmt.format(num_runs, min_iters, max_iters, mean_iters), end='')
 
@@ -834,15 +834,15 @@ class ModeSolver(object):
         idx1 = num_runs // 2
         idx2 = ((num_runs + 1) // 2) - 1
         median_iters = 0.5 * (sorted_iters[idx1] + sorted_iters[idx2])
-        if verbosity >= 2:
+        if verbosity.mpb >= 1:
             print(", median = {}".format(median_iters))
 
         mean_flops = self.eigensolver_flops / (num_runs * mean_iters)
-        if verbosity >= 2:
+        if verbosity.mpb >= 1:
             print("mean flops per iteration = {}".format(mean_flops))
 
         mean_time = self.total_run_time / (mean_iters * num_runs)
-        if verbosity >= 2:
+        if verbosity.mpb >= 1:
             print("mean time per iteration = {} s".format(mean_time))
 
     def _get_grid_size(self):
@@ -896,7 +896,7 @@ class ModeSolver(object):
 
         init_time = time.time()
 
-        if verbosity >= 2:
+        if verbosity.mpb >= 1:
             print("Initializing eigensolver data")
             print("Computing {} bands with {} tolerance".format(self.num_bands, self.tolerance))
 
@@ -905,7 +905,7 @@ class ModeSolver(object):
         if isinstance(reset_fields, basestring):
             self.load_eigenvectors(reset_fields)
 
-        if verbosity >= 2:
+        if verbosity.mpb >= 1:
             print("{} k-points".format(len(self.k_points)))
 
             for kp in self.k_points:
@@ -924,7 +924,7 @@ class ModeSolver(object):
                 solve_kpoint_time = time.time()
                 self.mode_solver.solve_kpoint(k)
                 self.iterations = self.mode_solver.get_iterations()
-                if verbosity >= 2:
+                if verbosity.mpb >= 1:
                     print("elapsed time for k point: {}".format(time.time() - solve_kpoint_time))
                 self.freqs = self.get_freqs()
                 self.all_freqs[i, :] = np.array(self.freqs)
@@ -952,12 +952,12 @@ class ModeSolver(object):
                 self.gap_list = []
 
         end = time.time() - start
-        if verbosity >= 2:
+        if verbosity.mpb >= 1:
             print("total elapsed time for run: {}".format(end))
         self.total_run_time += end
         self.eigensolver_flops = self.mode_solver.get_eigensolver_flops()
         self.parity = self.mode_solver.get_parity_string()
-        if verbosity >= 2:
+        if verbosity.mpb >= 1:
             print("done")
 
     def run(self, *band_functions):
@@ -1024,7 +1024,7 @@ class ModeSolver(object):
                 # First, look in the cached table
                 tab_val = bktab.get((b, k), None)
                 if tab_val:
-                    if verbosity >= 2:
+                    if verbosity.mpb >= 1:
                         print("find-k {} at {}: {} (cached)".format(b, k, tab_val[0]))
                     return tab_val
                 # Otherwise, compute bands and cache results
@@ -1046,7 +1046,7 @@ class ModeSolver(object):
                         bktab[(_b, k)] = (_f - omega, _v)
 
                     fun = self.freqs[-1] - omega
-                    if verbosity >= 2:
+                    if verbosity.mpb >= 1:
                         print("find-k {} at {}: {}".format(b, k, fun))
                     return (fun, v[-1])
 
@@ -1075,7 +1075,7 @@ class ModeSolver(object):
         self.num_bands = num_bands_save
         self.k_points = kpoints_save
         ks = list(reversed(ks))
-        if verbosity >= 2:
+        if verbosity.mpb >= 1:
             print("{}kvals:, {}, {}, {}".format(self.parity, omega, band_min, band_max), end='')
             for k in korig:
                 print(", {}".format(k), end='')
@@ -1239,7 +1239,7 @@ def output_dpwr_in_objects(output_func, min_energy, objects=[]):
         ms.get_dfield(which_band, False)
         ms.compute_field_energy()
         energy = ms.compute_energy_in_objects(objects)
-        if verbosity >= 2:
+        if verbosity.mpb >= 1:
             fmt = "dpwr:, {}, {}, {} "
             print(fmt.format(which_band, ms.freqs[which_band - 1], energy))
         if energy >= min_energy: