+++ python/visualization.py
@@ -1,7 +1,5 @@
 # -*- coding: utf-8 -*-
 from collections import namedtuple
-from collections import OrderedDict
-from collections import Sequence
 import warnings
 
 import numpy as np
@@ -846,7 +844,7 @@ class Animate2D(object):
 
             # Normalize the frames, if requested, and export
             if self.normalize and mp.am_master():
-                if mp.verbosity > 0:
+                if mp.verbosity.meep > 0:
                     print("Normalizing field data...")
                 fields = np.array(self.cumulative_fields) / np.max(np.abs(self.cumulative_fields),axis=(0,1,2))
                 for k in range(len(self.cumulative_fields)):
@@ -951,7 +949,7 @@ class Animate2D(object):
                         '-vf', 'pad=width=ceil(iw/2)*2:height=ceil(ih/2)*2',
                         '-an', filename  # output filename
             ]
-            if mp.verbosity > 0:
+            if mp.verbosity.meep > 0:
                 print("Generating GIF...")
             proc = Popen(command, stdin=PIPE, stdout=PIPE, stderr=PIPE)
             for i in range(len(self._saved_frames)):
@@ -988,7 +986,7 @@ class Animate2D(object):
                         '-vf', 'pad=width=ceil(iw/2)*2:height=ceil(ih/2)*2',
                         '-an', filename  # output filename
             ]
-            if mp.verbosity > 0:
+            if mp.verbosity.meep > 0:
                 print("Generating MP4...")
             proc = Popen(command, stdin=PIPE, stdout=PIPE, stderr=PIPE)
             for i in range(len(self._saved_frames)):